Fixed read speed chart, updated min/max/avg

+++ src/benchmark/hdd/UpdateChart.java
@@ -1,5 +1,5 @@
 package benchmark.hdd;
 
 public interface UpdateChart {
-	public void updateData(int x, int y);
+	public void updateData(double x, double y);
 }
